backend/requirements.txt

Flask==2.3.3
Flask-CORS==4.0.0
PyMySQL==1.1.0
cryptography==41.0.7
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
pandas==2.1.4
openpyxl==3.1.2
python-dotenv==1.0.0
playwright==1.58.0
